server/app/services/image_service.py
import os
from google.cloud import aiplatform
from google.oauth2 import service_account
import json
from dotenv import load_dotenv

load_dotenv()

# Initialize Google Cloud AI Platform (only if credentials are available)
credentials_path = os.getenv("GOOGLE_APPLICATION_CREDENTIALS")
if credentials_path and os.path.exists(credentials_path):
    try:
        credentials = service_account.Credentials.from_service_account_file(credentials_path)
        aiplatform.init(project=os.getenv("GOOGLE_CLOUD_PROJECT"), location="us-central1", credentials=credentials)
        print("Google Cloud AI Platform initialized successfully")
    except Exception as e:
        print(f"Failed to initialize Google Cloud: {e}")
        credentials = None
else:
    print("Google Cloud credentials not found - using mock image generation")
    credentials = None

async def generate_images(prompt: str, num_images: int = 4) -> list[str]:
    """
    Generate images using Google Imagen API or mock data
    """
    try:
        if credentials:
            # TODO: Implement actual Google Imagen API call
            # For now, return mock URLs
            pass
        
        # Mock image URLs (replace with real generated images)
        image_urls = [
            f"https://via.placeholder.com/512x512?text=Generated+Image+{i+1}"
            for i in range(num_images)
        ]
        
        return image_urls
        
        # Actual implementation would look like:
        # from vertexai.preview.vision_models import ImageGenerationModel
        # model = ImageGenerationModel.from_pretrained("imagegeneration@005")
        # response = model.generate_images(prompt=prompt, number_of_images=num_images)
        # return [image._as_dict()["uri"] for image in response]
        
    except Exception as e:
        raise Exception(f"Failed to generate images: {str(e)}")

def enhance_prompt(prompt: str) -> str:
    """
    Enhance user prompt using AI (placeholder)
    """
    # TODO: Implement prompt enhancement using OpenAI or similar
    return prompt
