server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
httpx==0.25.2
google-auth==2.23.4
google-cloud-aiplatform==1.36.4
python-dotenv==1.0.0
alembic==1.12.1
